requirements.txt
google-cloud-vision>=0.34.0
google-cloud-storage>=0.34.0
